+++ Sources/Core/AlulaCore/Bootstrap.swift
@@ -310,3 +310,5 @@ struct HealthTrackingService: Service {
         return logger
     }
 }
+
+extension LifecycleSettingsError: ModuleConfigurationError {}