+++ Sources/Core/AlulaCore/JSONLogHandler.swift
@@ -164,3 +164,5 @@ struct LoggingSettingsError: Error, CustomStringConvertible {
     let description: String
     init(_ description: String) { self.description = description }
 }
+
+extension LoggingSettingsError: ModuleConfigurationError {}