+++ Sources/Web/AlulaWeb/WebModule.swift
@@ -183,3 +183,5 @@ struct WebConfigurationError: Error, CustomStringConvertible {
     let description: String
     init(_ description: String) { self.description = description }
 }
+
+extension WebConfigurationError: ModuleConfigurationError {}